services/user-management/requirements-dev.txt
# Include production dependencies
-r requirements.txt

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
hypothesis>=6.92.0
moto>=4.2.0

# Code quality
flake8>=6.1.0
black>=23.12.0
mypy>=1.7.0
bandit>=1.7.5

# Local development
boto3-stubs[dynamodb,secretsmanager]>=1.34.0
